+++ sys/contrib/openzfs/tests/zfs-tests/tests/functional/history/history_007_pos.ksh
@@ -7,7 +7,7 @@
 # You may not use this file except in compliance with the License.
 #
 # You can obtain a copy of the license at usr/src/OPENSOLARIS.LICENSE
-# or https://opensource.org/licenses/CDDL-1.0.
+# or http://www.opensolaris.org/os/licensing.
 # See the License for the specific language governing permissions
 # and limitations under the License.
 #
@@ -49,7 +49,7 @@ function cleanup
 	poolexists $migratedpoolname &&  \
 		log_must zpool destroy -f $migratedpoolname
 
-	rm -rf $import_dir
+	[[ -d $import_dir ]] && rm -rf $import_dir
 }
 
 log_assert "Verify command history moves with migrated pool."
@@ -62,7 +62,7 @@ migratedpoolname=$MIGRATEDPOOLNAME
 typeset -i RET=1
 typeset -i linenum=0
 
-[[ ! -d $import_dir ]] && log_must mkdir -p $import_dir
+[[ ! -d $import_dir ]] && log_must mkdir $import_dir
 
 # We test the migrations on both uniform platform and cross platform
 for arch in "i386" "sparc"; do
@@ -70,7 +70,7 @@ for arch in "i386" "sparc"; do
 	orig_cmds_f=$import_dir/${arch}.orig_history.txt
 	# remove blank line
 	orig_cmds_f1=$import_dir/${arch}.orig_history_1.txt
-	grep -v "^$" $orig_cmds_f > $orig_cmds_f1
+	cat $orig_cmds_f | grep -v "^$" > $orig_cmds_f1
 
 	log_must cp $tst_dir/${arch}.migratedpool.DAT.Z $import_dir
 	log_must uncompress -f $import_dir/${arch}.migratedpool.DAT.Z
@@ -80,7 +80,10 @@ for arch in "i386" "sparc"; do
 	    log_must zpool destroy -f $migratedpoolname
 
 	log_must zpool import -d $import_dir $migratedpoolname
-	log_must eval "TZ=$TIMEZONE zpool history $migratedpoolname | grep -v \"^\$\" >$migrated_cmds_f"
+	TZ=$TIMEZONE zpool history $migratedpoolname | grep -v "^$" \
+	    >$migrated_cmds_f
+	RET=$?
+	(( $RET != 0 )) && log_fail "zpool history $migratedpoolname fails."
 
 	# The migrated history file should differ with original history file on
 	# two commands -- 'export' and 'import', which are included in migrated
@@ -89,20 +92,21 @@ for arch in "i386" "sparc"; do
 	# then compare this filtered file with the original history file. They
 	# should be identical at this time.
 	for subcmd in "export" "import"; do
-		grep -q "$subcmd" $migrated_cmds_f ||
-			log_fail "zpool $subcmd is not logged for" \
-			    "the imported pool $migratedpoolname."
+		grep "$subcmd" $migrated_cmds_f >/dev/null 2>&1
+		RET=$?
+		(( $RET != 0 )) && log_fail "zpool $subcmd is not logged for" \
+		    "the imported pool $migratedpoolname."
 	done
 
 	tmpfile=$import_dir/cmds_tmp.$$
-	linenum=$(wc -l < $migrated_cmds_f)
+	linenum=`cat $migrated_cmds_f | wc -l`
 	(( linenum = linenum - 2 ))
 	head -n $linenum $migrated_cmds_f > $tmpfile
 	log_must diff $tmpfile $orig_cmds_f1
 
 	# cleanup for next loop testing
 	log_must zpool destroy -f $migratedpoolname
-	log_must rm -f $(ls $import_dir)
+	log_must rm -f `ls $import_dir`
 done
 
 log_pass "Verify command history moves with migrated pool."
